drivers: video: ar0144: temporary workaround
The RTL does not support RGGB8, but only BGGR8, so the RTL driver will
configure BGGR8 instead.

Temporary workaround until we understand how the debayer core is expected
to support, to lie about format supported to at least get an image out.

Once the image comes out, decide of a solution for this issue.